Lots
Added:
InputGen.py - Script that generates a text file as described in the outline. It doesn't generate -1 values because I'm not sure how to only generate possible solutions, but the rest of the code accounts for -1 values.

HexGrid.py -  defines the HexGrid properties as well as the Node properties

Problemsearch.py - A generic implementation of the graph_search algorithm. Since it is generic, it is easy to swap out search heuristics.

searchstrategies.py - Home for search heuristics. I implemented the basic Djisktra's algorithm but theres definitely a heuristic for shortest distance on a hex lattice if anyone wants to implement that.

explored.py - Generic hash set to manage any way you want to define a problem state.

queues.py - PriorityQueue that takes in a function with which to sort the elements.

driver.py - Input the path to the input text file here. It uses Rick's code to interpret the input text file and generate a list of node values. It also prints the time it takes to find the shortest path, currently the slowest runtime I've found has been fractions of seconds.